myap/examapp/requirements.txt
asgiref==3.7.2
contourpy==1.0.7
cycler==0.11.0
Django==4.2.2
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
gunicorn==20.1.0